requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.2
dill==0.3.8
distlib==0.3.8
fastapi==0.109.0
filelock==3.13.1
flake8==7.0.0
future-fstrings==1.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
networkx==3.2.1
nodeenv==1.8.0
outcome==1.3.0.post0
packaging==23.2
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
pylint==3.0.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-depends==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomlkit==0.12.3
trio==0.24.0
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240106
types-requests==2.31.0.20240125
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
uvicorn==0.26.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
